implemented main.py and tested;

+++ calculator/operations.py
@@ -20,5 +20,5 @@ def sigma(n: int) -> int:
     """Sum of n numbers"""
     if n < 0:
         raise ValueError("Please enter an integer > 0")
-    return n*(n+1)/2
+    return int(n*(n+1)/2)
